moved all components to components folder
