+++ archive/App.tsx
@@ -1,6 +1,6 @@
 // import { Counter } from './features/counter/Counter'
 import React from 'react';
-import MainPage from './features/mainPage/mainPageContainer';
+import MainPage from '../src/components/App';
 import './stylesheets/styles.scss';
 
 export default function App (): JSX.Element {
